# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Xuwe API
  version: 0.1.0
  description: 基于 OIDC Bearer token 与本地 PostgreSQL RBAC 的管理 API。
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /roles/{role_id}:
    parameters:
      - $ref: "#/components/parameters/RoleId"
    get:
      operationId: getRole
      summary: 返回角色及其权限
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 角色详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateRole
      summary: 修改自定义角色元数据
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRole"
      responses:
        "200":
          description: 更新后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    delete:
      operationId: deleteRole
      summary: 删除未被引用的自定义角色
      security:
        - bearerAuth: []
      responses:
        "204":
          description: 角色已删除
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RoleId:
      name: role_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Bearer token 缺失、无效或已过期
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: 用户已停用或缺少权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: 资源冲突、操作会移除最后一个管理员，或试图修改内置超级管理员
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationFailed:
      description: 字段校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    Permission:
      type: object
      required: [id, key, name]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          pattern: "^[a-z][a-z0-9._-]+:[a-z][a-z0-9._-]+$"
        name:
          type: string
        description:
          type: [string, "null"]
    Role:
      type: object
      required: [id, key, name, is_system, permissions, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type: [string, "null"]
        is_system:
          type: boolean
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        created_at:
          type: integer
          format: int64
          description: 角色创建时间的 Unix 秒时间戳
        updated_at:
          type: integer
          format: int64
          description: 角色最后更新时间的 Unix 秒时间戳
...
    UpdateRole:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, "null"]
          maxLength: 1000
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string